+++ CHANGES.rst
@@ -10,6 +10,10 @@
     - plotting/test_qc_filter
     - general
     - tools/clustering
+    - tools/bam
+    - highly_variable
+    - tables
+    - utils/general
 
 0.7.0 (23-04-24)
 -------------